requirements.txt
Django
djangorestframework
psycopg2
flake8
djoser
django-extensions